% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBrIII.R
\name{dBrIII}
\alias{dBrIII}
\title{Probability density function of BrIII distribution}
\usage{
dBrIII(x, para = c(1, 2, 0.5))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(scale, shape1, shape2)}
}
\value{
Probability density function
}
\description{
Probability density function of BrIII distribution
}
\examples{

d <- dBrIII(x = 108.4992, para = c(10, 0.25, 0.5))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
