% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgam.R
\name{dgam}
\alias{dgam}
\title{Probability density function of Gamma distribution}
\usage{
dgam(x, para = c(1, 2, 0.5))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(shape, scale)}
}
\value{
Probability density function
}
\description{
Probability density function of Gamma distribution
}
\examples{

d <- dgam(x = 0.1, para = c(0.1, 0.2))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
