% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgev.R
\name{dgev}
\alias{dgev}
\title{Probability density function of GEV distribution}
\usage{
dgev(x, para)
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(location, scale, shape)}
}
\value{
Probability density function
}
\description{
Probability density function of GEV distribution
}
\examples{

d <- dgev(x = 108.4992, para = c(10, 1, 1))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
