% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGG.R
\name{dGG}
\alias{dGG}
\title{Probability density function of Generalized Gamma (GG) distribution}
\usage{
dGG(x, para = c(10, 0.25, 0.5))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(scale, shape1, shape2)}
}
\value{
Probability density function
}
\description{
Probability density function of Generalized Gamma (GG) distribution
}
\examples{

d <- dGG(x = 108.4992, para = c(10, 0.25, 0.5))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
