% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgno.R
\name{dgno}
\alias{dgno}
\title{Probability density function of Generalized normal Distribution}
\usage{
dgno(x, para = c(1, 2, 0.5))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(location, scale, shape)}
}
\value{
Probability density function
}
\description{
Probability density function of Generalized normal Distribution
}
\examples{

d <- dgno(x = 0.1, para = c(1, 2, 0.5))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
