% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgpa.R
\name{dgpa}
\alias{dgpa}
\title{Probability density function of Generalized Pareto Distribution}
\usage{
dgpa(x, para)
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(location, scale, shape)}
}
\value{
Probability density function
}
\description{
Probability density function of Generalized Pareto Distribution
}
\examples{

d <- dgpa(x = 0.1, para = c(1, 2, 0.5))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
