% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpe3.R
\name{dpe3}
\alias{dpe3}
\title{Probability density function of Pearson type-3 Distribution}
\usage{
dpe3(x, para = c(10, 1, 1.5))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(mu, sigma, gamma) that is c(location, scale, shape).}
}
\value{
Probability density function
}
\description{
Probability density function of Pearson type-3 Distribution
}
\examples{

d <- dpe3(x = 12, para = c(10, 1, 1.5))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
