% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_BrIII.R
\name{fit_BrIII}
\alias{fit_BrIII}
\title{Fit Burr Type-III (BrIII) Distribution}
\usage{
fit_BrIII(sl1, st2, st3)
}
\arguments{
\item{sl1}{1st l-moments}

\item{st2}{2nd l-moment ratio}

\item{st3}{3rd l-moment ratio}
}
\value{
A dataframe containing the scale parameter, the shape1 parameter, the shape2 parameter, the squared error of scale parameter, and the squared error of shape parameter
}
\description{
Fit Burr Type-III (BrIII) Distribution
}
\examples{

BrIII_par_valid <- fit_BrIII(sl1 = 10, st2 = 0.25, st3 = 0.1)
BrIII_par_invalid <- fit_BrIII(sl1 = 10, st2 = 0.5, st3 = 0.8)

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
