% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_BrXII.R
\name{fit_BrXII}
\alias{fit_BrXII}
\title{Fit Burr Type-XII (BrXII) Distribution}
\usage{
fit_BrXII(sl1, st2, st3)
}
\arguments{
\item{sl1}{1st l-moments}

\item{st2}{2nd l-moment ratio}

\item{st3}{3rd l-moment ratio}
}
\value{
A dataframe containing the scale parameter, the shape1 parameter, the shape2 parameter, the squared error of the scale parameter, and the squared error of the shape parameters.
}
\description{
Fit Burr Type-XII (BrXII) Distribution
}
\examples{

BrXII_par_valid <- fit_BrXII(sl1 = 10, st2 = 0.25, st3 = 0.25)
BrXII_par_invalid <- fit_BrXII(sl1 = 10, st2 = 0.5, st3 = 0.8)

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
