% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{fit_gam}
\alias{fit_gam}
\title{Fit Gamma distribution using the 'lmom' package}
\usage{
fit_gam(sl1, sl2, st3, st4)
}
\arguments{
\item{sl1}{sample 1st l-moment}

\item{sl2}{sample 2nd l-moment}

\item{st3}{sample 3rd l-moment ratio}

\item{st4}{sample 4th l-moment ratio}
}
\value{
A vector of parameters as alpha (shape) and beta (scale).
}
\description{
Fit Gamma distribution using the 'lmom' package
}
\examples{

gam_par <- fit_gam(15, 1.7, 0.04, -0.02)

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
