% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gev.R
\name{fit_gev}
\alias{fit_gev}
\title{Fit GEV distribution}
\usage{
fit_gev(sl1, sl2, st3)
}
\arguments{
\item{sl1}{sample 1st l-moment}

\item{sl2}{sample 2nd l-moment}

\item{st3}{sample 3rd l-moment ratio}
}
\value{
A dataframe containing the location parameter, the scale parameter, the shape parameter, and the squared error of shape parameters.
}
\description{
Fit GEV distribution
}
\examples{

GEV_par <- fit_gev(sl1 = 10, sl2 = 0.5, st3 = 0.8)

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
