% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_GG.R
\name{fit_GG}
\alias{fit_GG}
\title{Fit Generalized Gamma (GG) Distribution}
\usage{
fit_GG(sl1, st2, st3)
}
\arguments{
\item{sl1}{1st l-moments}

\item{st2}{2nd l-moment ratio}

\item{st3}{3rd l-moment ratio}
}
\value{
A dataframe containing the scale parameter, the shape1 parameter, the shape2 parameter, the squared error of scale parameter, and the squared error of shape parameters.
}
\description{
Fit Generalized Gamma (GG) Distribution
}
\examples{

GG_par_valid <- fit_GG(sl1 = 10, st2 = 0.4, st3 = 0.2)
GG_par_invalid <- fit_GG(sl1 = 1, st2 = 0.25, st3 = 0.25)

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
