% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FLOW_AMAX}
\alias{FLOW_AMAX}
\title{Annual maximum flow data at Water Survey of Canada WSC flow gauge number 08NA002 in BC, Vancouver, Canada. Lat: 51°14'36.8¨ N, Long: 116°54'46.6¨ W.}
\format{
A vector of observations of length equal to 112
\describe{
  \item{flow}{annual maximum flow observed per each year at one site}
}
}
\source{
coded in data-raw
}
\usage{
FLOW_AMAX
}
\description{
Annual maximum flow data at Water Survey of Canada WSC flow gauge number 08NA002 in BC, Vancouver, Canada. Lat: 51°14'36.8¨ N, Long: 116°54'46.6¨ W.
}
\keyword{datasets}
