% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FLOW_AMAX_MULT}
\alias{FLOW_AMAX_MULT}
\title{Annual maximum flow data at 10 hypothetical flow gauge.}
\format{
A data frame with 112 rows and 10 variables:
\describe{
  \item{flow_st1}{annual maximum flow observed per each year at site 1}
  \item{flow_st2}{annual maximum flow observed per each year at site 2}
  \item{flow_st3}{annual maximum flow observed per each year at site 3}
  \item{flow_st4}{annual maximum flow observed per each year at site 4}
  \item{flow_st5}{annual maximum flow observed per each year at site 5}
  \item{flow_st6}{annual maximum flow observed per each year at site 6}
  \item{flow_st7}{annual maximum flow observed per each year at site 7}
  \item{flow_st8}{annual maximum flow observed per each year at site 8}
  \item{flow_st9}{annual maximum flow observed per each year at site 9}
  \item{flow_st10}{annual maximum flow observed per each year at site 10}
}
}
\source{
coded in data-raw
}
\usage{
FLOW_AMAX_MULT
}
\description{
Annual maximum flow data at 10 hypothetical flow gauge.
}
\keyword{datasets}
