% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_julian.R
\name{get_julian}
\alias{get_julian}
\title{Get julian date from the begining of the year}
\usage{
get_julian(x)
}
\arguments{
\item{x}{date or a series of dates such as, as.Date("yyyy-mm-dd")}
}
\value{
A julian date between 1 and 365, note that in leap years the day 366 is considered as 365
}
\description{
Get julian date from the begining of the year
}
\examples{

get_julian(x = as.Date("1979-01-15"))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
