% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_lmom.R
\name{get_sample_lmom}
\alias{get_sample_lmom}
\title{Estimate sample L-moments and L-moment ratios}
\usage{
get_sample_lmom(x)
}
\arguments{
\item{x}{a series of quantiles}
}
\value{
A dataframe containing the 1st l-moment, the 2nd l-moment, the 3rd l-moment, the 4th l-moment, the 2nd l-moment ratio "L-variation", the 3rd l-moment ratio "L-skewness", and the 4th l-moment ratio "L-kurtosis"
}
\description{
Estimate sample L-moments and L-moment ratios
}
\examples{

sample_lmom <- get_sample_lmom((rnorm(n = 500, mean = 10, sd = 0.5)))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
