% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lspace_BrIII}
\alias{lspace_BrIII}
\title{L-space of Burr Type-III Distribution (BrIII)}
\format{
A ggplot 
\describe{
  \item{data}{}
  \item{layers}{}
  \item{scales}{}
  \item{mapping}{}
  \item{theme}{}
  \item{coordinates}{}
  \item{facet}{}
  \item{plot_env}{}
  \item{labels}{}
}
}
\source{
coded in data-raw
}
\usage{
lspace_BrIII
}
\description{
This is a plot of the L-space of BrIII Distribution with L-variation on x-axis and L-skewness on y-axis.
The L-space is bounded by shape1 in the range of 0.01 to 150.01, and by shape2 in the range of 0.005 to 0.999.
}
\keyword{datasets}
