% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lspace_BrIII.xy}
\alias{lspace_BrIII.xy}
\title{coordinates of the L-space of Burr Type-III Distribution (BrIII)}
\format{
A ggplot 
\describe{
  \item{x}{l-variatoin "t2"}
  \item{y}{l-skewness "t3"}
}
}
\source{
coded in data-raw
}
\usage{
lspace_BrIII.xy
}
\description{
This is a plot of the L-space of BrIII Distribution with L-variation on x-axis and L-skewness on y-axis.
The L-space is bounded by shape1 in the range of 0.01 to 150.01, and by shape2 in the range of 0.005 to 0.999.
}
\keyword{datasets}
