% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lspace_BrXII}
\alias{lspace_BrXII}
\title{L-space of Burr Type-XII Distribution (BrXII)}
\format{
A ggplot 
\describe{
  \item{data}{}
  \item{layers}{}
  \item{scales}{}
  \item{mapping}{}
  \item{theme}{}
  \item{coordinates}{}
  \item{facet}{}
  \item{plot_env}{}
  \item{labels}{}
}
}
\source{
coded in data-raw
}
\usage{
lspace_BrXII
}
\description{
This is a plot of the L-space of BrXII Distribution with L-variation on x-axis and L-skewness on y-axis.
The L-space is bounded by shape1 in the range of 0.1 to 150, and by shape2 in the range of 0.001 to 1.
}
\keyword{datasets}
