% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lspace_GG.xy}
\alias{lspace_GG.xy}
\title{coordinates of the L-space of Generalized Gamma Distribution (GG)}
\format{
A ggplot 
\describe{
  \item{x}{l-variatoin "t2"}
  \item{y}{l-skewness "t3"}
}
}
\source{
coded in data-raw
}
\usage{
lspace_GG.xy
}
\description{
This is a plot of the L-space of GG Distribution with L-variation on x-axis and L-skewness on y-axis.
The L-space is bounded by shape1 in the range of 0.1 to 5.9, and by shape2 in the range of 0.19 to 38.
}
\keyword{datasets}
