% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgam.R
\name{pgam}
\alias{pgam}
\title{Cumulative distribution function of Gamma distribution}
\usage{
pgam(x, para = c(1.5, 1))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(shape, scale)}
}
\value{
Non-exceedance probability from the cumulative distribution function.
}
\description{
Cumulative distribution function of Gamma distribution
}
\examples{

u <- pgam(x = 0.1, para = c(0.1, 0.2))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
