% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pGG.R
\name{pGG}
\alias{pGG}
\title{Cumulative distribution function of Generalized Gamma (GG) distribution}
\usage{
pGG(x, para = c(10, 0.25, 0.5))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(scale, shape1, shape2)}
}
\value{
Non-exceedance probability from the cumulative distribution function.
}
\description{
Cumulative distribution function of Generalized Gamma (GG) distribution
}
\examples{

u <- pGG(x = 108.4992, para = c(10, 0.25, 0.5))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
