% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgno.R
\name{pgno}
\alias{pgno}
\title{Cumulative distribution function of Generalized Normal Distribution}
\usage{
pgno(x, para = c(10, 1.5, 1))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(location, scale, shape)}
}
\value{
Non-exceedance probability from the cumulative distribution function.
}
\description{
Cumulative distribution function of Generalized Normal Distribution
}
\examples{

u <- pgno(x = 10.1, para = c(10, 0.1, 0.2))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
