% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgpa.R
\name{pgpa}
\alias{pgpa}
\title{Cumulative distribution function of Generalized Pareto Distribution}
\usage{
pgpa(x, para = c(1, 1, 1))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(location, scale, shape)}
}
\value{
Non-exceedance probability from the cumulative distribution function.
}
\description{
Cumulative distribution function of Generalized Pareto Distribution
}
\examples{

u <- pgpa(x = 1.2, para = c(1, 2, 0.5))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
