% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pln3.R
\name{pln3}
\alias{pln3}
\title{Cumulative distribution function of Lognormal-3 Distribution}
\usage{
pln3(x, para = c(0, 0, 1))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(zeta, mu, sigma) that is c(lower bound, mean on log scale, standard deviation on log scale).}
}
\value{
Non-exceedance probability from the cumulative distribution function.
}
\description{
Cumulative distribution function of Lognormal-3 Distribution
}
\examples{

u <- pln3(x = 12, para = c(0, 0, 1))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
