% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnor.R
\name{pnor}
\alias{pnor}
\title{Cumulative distribution function of Noramal Distribution}
\usage{
pnor(x, para = c(10, 1.5))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(location, scale)}
}
\value{
Non-exceedance probability from the cumulative distribution function.
}
\description{
Cumulative distribution function of Noramal Distribution
}
\examples{

u <- pnor(x = 11, para = c(10, 1.5))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
