% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppe3.R
\name{ppe3}
\alias{ppe3}
\title{Cumulative distribution function of Pearson type-3 Distribution}
\usage{
ppe3(x, para = c(10, 1, 1.5))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(mu, sigma, gamma) that are c(location, scale, shape).}
}
\value{
Non-exceedance probability from the cumulative distribution function.
}
\description{
Cumulative distribution function of Pearson type-3 Distribution
}
\examples{

u <- ppe3(x = 12, para = c(10, 1, 1.5))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
