% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgam.R
\name{tgam}
\alias{tgam}
\title{Return period function of Gamma distribution}
\usage{
tgam(x, para = c(1.5, 1))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(shape, scale)}
}
\value{
Return Period/s corresponding to quantile/s.
}
\description{
Return period function of Gamma distribution
}
\examples{

RP <- tgam(x = 0.1, para = c(0.1, 0.2))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
