% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgev.R
\name{tgev}
\alias{tgev}
\title{Return period function of GEV distribution}
\usage{
tgev(x, para)
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(location, scale, shape)}
}
\value{
Return Period/s corresponding to quantile/s.
}
\description{
Return period function of GEV distribution
}
\examples{

RP <- tgev(x = 108.4992, para = c(10, 1, 1))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
