% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tGG.R
\name{tGG}
\alias{tGG}
\title{Return period function of Generalized Gamma distribution}
\usage{
tGG(x, para = c(10, 0.25, 0.5))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(scale, shape1, shape2)}
}
\value{
Return Period/s corresponding to quantile/s.
}
\description{
Return period function of Generalized Gamma distribution
}
\examples{

RP <- tGG(x = 108.4992, para = c(10, 0.25, 0.5))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
