% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgpa.R
\name{tgpa}
\alias{tgpa}
\title{Return period function of Generalized Pareto distribution}
\usage{
tgpa(x, para = c(1, 1, 1))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(location, scale, shape)}
}
\value{
Return Period/s corresponding to quantile/s.
}
\description{
Return period function of Generalized Pareto distribution
}
\examples{

RP <- tgpa(x = 1.2, para = c(1, 2, 0.5))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
