% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tln3.R
\name{tln3}
\alias{tln3}
\title{Return period function of Lognormal-3 distribution}
\usage{
tln3(x, para = c(0, 0, 1))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(zeta, mu, sigma) that is c(lower bound, mean on log scale, standard deviation on log scale).}
}
\value{
Return Period/s corresponding to quantile/s.
}
\description{
Return period function of Lognormal-3 distribution
}
\examples{

RP <- tln3(x = 12, para = c(0, 0, 1))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
