% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnor.R
\name{tnor}
\alias{tnor}
\title{Return period function of Noramal distribution}
\usage{
tnor(x, para = c(10, 1.5))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(location, scale)}
}
\value{
Return Period/s corresponding to quantile/s.
}
\description{
Return period function of Noramal distribution
}
\examples{

RP <- tnor(x = 11, para = c(10, 1.5))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
