% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpe3.R
\name{tpe3}
\alias{tpe3}
\title{Return period function of Pearson type-3 distribution}
\usage{
tpe3(x, para = c(10, 1, 1.5))
}
\arguments{
\item{x}{quantile/s}

\item{para}{parameters as c(mu, sigma, gamma) that are c(location, scale, shape).}
}
\value{
Return Period/s corresponding to quantile/s.
}
\description{
Return period function of Pearson type-3 distribution
}
\examples{

RP <- tpe3(x = 12, para = c(10, 1, 1.5))

}
\author{
Mohanad Zaghloul [aut, cre], Simon Michael Papalexiou [aut, ths], Amin Elshorbagy [aut, ths]
}
